Allow searching for credential details
