+++ frontend/src/components/Credentials/GCredentialDetailsItemContent.vue
@@ -34,9 +34,7 @@ SPDX-License-Identifier: Apache-2.0
 </template>
 
 <script>
-import { decodeBase64 } from '@/utils'
-
-import get from 'lodash/get'
+import { secretDetails } from '@/composables/credential/helper'
 
 export default {
 
@@ -67,7 +65,7 @@ export default {
           },
         ]
       } else if (this.credential?.kind === 'Secret') {
-        const details = this.getCredentialDetails(this.credential)
+        const details = secretDetails(this.credential, this.providerType)
         if (details) {
           return details
         }
@@ -87,176 +85,6 @@ export default {
           disabledText: true,
         },
       ]
-    },
-  },
-  methods: {
-    getCredentialDetails (secret) {
-      const secretData = secret.data || {}
-      const getGCPProjectId = () => {
-        const serviceAccount = get(secretData, ['serviceaccount.json'])
-        return get(JSON.parse(decodeBase64(serviceAccount)), ['project_id'])
-      }
-      try {
-        switch (this.providerType) {
-          // infra
-          case 'openstack':
-            return [
-              {
-                label: 'Domain Name',
-                value: decodeBase64(secretData.domainName),
-              },
-              {
-                label: 'Tenant Name',
-                value: decodeBase64(secretData.tenantName),
-              },
-            ]
-          case 'vsphere':
-            return [
-              {
-                label: 'vSphere Username',
-                value: decodeBase64(secretData.vsphereUsername),
-              },
-              {
-                label: 'NSX-T Username',
-                value: decodeBase64(secretData.nsxtUsername),
-              },
-            ]
-          case 'aws':
-            return [
-              {
-                label: 'Access Key ID',
-                value: decodeBase64(secretData.accessKeyID),
-              },
-            ]
-          case 'azure':
-            return [
-              {
-                label: 'Subscription ID',
-                value: decodeBase64(secretData.subscriptionID),
-              },
-            ]
-          case 'gcp':
-            return [
-              {
-                label: 'Project',
-                value: getGCPProjectId(),
-              },
-            ]
-          case 'alicloud':
-            return [
-              {
-                label: 'Access Key ID',
-                value: decodeBase64(secretData.accessKeyID),
-              },
-            ]
-          case 'metal':
-            return [
-              {
-                label: 'API URL',
-                value: decodeBase64(secretData.metalAPIURL),
-              },
-            ]
-          case 'hcloud':
-            return [
-              {
-                label: 'Hetzner Cloud Token',
-                hidden: true,
-              },
-            ]
-          case 'openstack-designate':
-            return [
-              {
-                label: 'Domain Name',
-                value: decodeBase64(secretData.domainName),
-              },
-              {
-                label: 'Tenant Name',
-                value: decodeBase64(secretData.tenantName),
-              },
-            ]
-          // dns
-          case 'aws-route53':
-            return [
-              {
-                label: 'Access Key ID',
-                value: decodeBase64(secretData.accessKeyID),
-              },
-            ]
-          case 'azure-dns':
-          case 'azure-private-dns':
-            return [
-              {
-                label: 'Subscription ID',
-                value: decodeBase64(secretData.subscriptionID),
-              },
-            ]
-          case 'google-clouddns':
-            return [
-              {
-                label: 'Project',
-                value: decodeBase64(secretData.project),
-              },
-            ]
-          case 'alicloud-dns':
-            return [
-              {
-                label: 'Access Key ID',
-                value: decodeBase64(secretData.accessKeyID),
-              },
-            ]
-          case 'infoblox-dns':
-            return [
-              {
-                label: 'Infoblox Username',
-                value: decodeBase64(secretData.USERNAME),
-              },
-            ]
-          case 'cloudflare-dns':
-            return [
-              {
-                label: 'API Key',
-                hidden: true,
-              },
-            ]
-          case 'netlify-dns':
-            return [
-              {
-                label: 'API Key',
-                hidden: true,
-              },
-            ]
-          case 'rfc2136':
-            return [
-              {
-                label: 'Server',
-                value: decodeBase64(secretData.Server),
-              },
-              {
-                label: 'TSIG Key Name',
-                value: decodeBase64(secretData.TSIGKeyName),
-              },
-              {
-                label: 'Zone',
-                value: decodeBase64(secretData.Zone),
-              },
-            ]
-          case 'powerdns':
-            return [
-              {
-                label: 'Server',
-                value: decodeBase64(secretData.server),
-              },
-              {
-                label: 'API Key',
-                hidden: true,
-              },
-            ]
-          default:
-            return undefined
-        }
-      } catch (err) {
-        return undefined
-      }
     },
   },
 }
